separate TestRawArray
