+++ tests/chainer_tests/training_tests/updaters_tests/test_multiprocess_parallel_updater.py
@@ -129,37 +129,6 @@ class TestGatherScatter(unittest.TestCase):
             mpu.gather_params(model)
 
 
-class SimpleNetRawArray(chainer.Chain):
-
-    def __init__(self, testcase):
-        super(SimpleNetRawArray, self).__init__()
-        with self.init_scope():
-            self.conv = chainer.links.Convolution2D(2, 2, 3)
-            self.fc = chainer.links.Linear(18, 2)
-
-        self.train = True
-        self.call_called = 0
-        self.testcase = testcase
-
-    def clear(self):
-        self.loss = None
-        self.accuracy = None
-
-    def __call__(self, x, t):
-        self.testcase.assertNotIsInstance(x, chainer.Variable)
-        self.testcase.assertNotIsInstance(t, chainer.Variable)
-
-        self.call_called += 1
-
-        h = chainer.functions.relu(self.conv(x))
-        y = self.fc(h)
-
-        self.loss = chainer.functions.softmax_cross_entropy(y, t)
-        self.accuracy = chainer.functions.accuracy(y, t)
-
-        return self.loss
-
-
 def _run_test_snippet(name, *args):
     script_path = os.path.join(
         os.path.dirname(__file__), 'snippets/{}'.format(name))
@@ -178,26 +147,13 @@ class TestRawArray(unittest.TestCase):
         pass
 
     @attr.gpu
+    @unittest.skipUnless(mpu.MultiprocessParallelUpdater.available(),
+                         'MultiprocessParallelUpdater is not available.')
     def test_update_uses_raw_array(self):
-        if mpu.MultiprocessParallelUpdater.available():
-            model = SimpleNetRawArray(self)
-            dataset = [((numpy.ones((2, 5, 5)) * i).astype(numpy.float32),
-                        numpy.int32(0)) for i in range(100)]
-
-            batch_size = 5
-            devices = (0,)
-            iters = [chainer.iterators.SerialIterator(i, batch_size) for i in
-                     chainer.datasets.split_dataset_n_random(
-                         dataset, len(devices))]
-            optimizer = chainer.optimizers.SGD(lr=1.0)
-            optimizer.setup(model)
-
-            with testing.assert_warns(UserWarning):
-                updater = mpu.MultiprocessParallelUpdater(
-                    iters, optimizer, devices=devices)
-            updater.update()
-
-            self.assertEqual(model.call_called, 1)
+        ret, stdoutdata, stderrdata = _run_test_snippet('raw_array.py')
+        assert ret == 0, (
+            '[stdout]:{!r}\n'
+            '[stderr]:{!r}'.format(stdoutdata, stderrdata))
 
 
 class TestChildReporter(unittest.TestCase):
